refactor: remove season correlation pattern and specials detection logic
- Deleted PatternDetector class responsible for identifying mapping patterns between Western and anime seasons.
- Removed SpecialsDetector class that analyzed and categorized special episodes.
- Eliminated related types and interfaces for enhanced models, mapping patterns, mapping rules, provenance tracking, and specials handling.
- Updated series repository and transformer imports to reflect the removal of obsolete files.
- Cleaned up utility functions and types related to series episodes and seasons.
