+++ src/episodes/repository/helpers/qualifier.ts
@@ -1,0 +1,3 @@
+import { Format } from '../../../service/notify/transformer/enums.ts';
+
+export const isMovie = (format?: Format): boolean => format == Format.MOVIE;
