+++ src/series/repository/helpers/qualifier.ts
@@ -1,0 +1,3 @@
+import { MalType } from '../../../service/jikan/remote/index.ts';
+
+export const isManga = (type?: MalType): boolean => type == MalType.Manga;
