+++ src/series/transformer/index.ts
@@ -1,3 +1,1 @@
-export * from './season.transformer.ts';
 export * from './series.transformer.ts';
-export * from './types.ts';
