+++ src/series/types.ts
@@ -69,70 +69,6 @@ export type SeriesImageAttributes = {
   type: 'BACKDROP' | 'POSTER' | 'LOGO';
 };
 
-// Full image sets used for top-level media image
-export type SeriesImage = {
-  backdrops: SeriesImageAttributes[];
-  logos: SeriesImageAttributes[];
-  posters: SeriesImageAttributes[];
-};
-
-export type SeriesImageSimple = {
-  extraLarge?: string;
-  large?: string;
-  medium?: string;
-  banner?: string | null;
-  logo?: string | null;
-};
-
-export interface SeriesEpisodeCrew {
-  job?: string;
-  department?: string;
-  creditId: string;
-  adult?: boolean;
-  id: number;
-  knownFor: string;
-  name: string;
-  originalName: string;
-  popularity: number;
-  image?: string;
-  character?: string;
-  order?: number;
-}
-
-export type SeriesEpisode = {
-  id: number;
-  tvdbShowId: number;
-  tvdbId: number;
-  seasonNumber: number;
-  episodeNumber: number;
-  absoluteEpisodeNumber?: number;
-  airedBeforeSeasonNumber?: number;
-  airedBeforeEpisodeNumber?: number;
-  airedAfterSeasonNumber?: number;
-  airedAfterEpisodeNumber?: number;
-  title?: string;
-  airDate: Instant;
-  runtime?: number;
-  overview?: string;
-  image?: string;
-  name?: string;
-  poster?: string;
-  crew: SeriesEpisodeCrew[];
-  guests: SeriesEpisodeCrew[];
-};
-
-export type SeriesSeason = {
-  tmdbId: number;
-  airDate: Instant;
-  episodeCount: number;
-  name: string;
-  overview: string;
-  number: number;
-  cover?: string;
-  image: SeriesImageSimple;
-  episodes: SeriesEpisode[];
-};
-
 export type SeriesTrailer = {
   id: string;
   site: string;
